test/requirements.txt
locust
websockets
websocket-client